Add files via upload
The turtle module is an extended reimplementation of the same-named module from the Python standard distribution up to version Python 2.5.
Turtle! is a Windows app that lets kids drag and drop commands to create their own programs. In the process, they learn to solve problems, design projects, and express themselves creatively on the computer.
Turtle() is the constructor method of the class Turtle ; it returns an instance of the class. If you don't assign the output to a variable, it basically creates an instance that is immediately discarded (that's not entirely true, as you could still access it with the special python variable _ right after that call.)